cambio de tipo de dato cedula y telefono

+++ wowguaupaseador/app/src/main/java/co/edu/javeriana/wowguau_paseador/model/Usuario.java
@@ -12,9 +12,9 @@ import java.util.Date;
 public abstract class Usuario implements Serializable {
     private String correo;
     private String nombre;
-    private int cedula;
+    private long cedula;
     private Date fechaNacimiento;
-    private int telefono;
+    private long telefono;
     private String genero;
     private Direccion direccion;
     private float saldo;
@@ -23,7 +23,7 @@ public abstract class Usuario implements Serializable {
     @Exclude
     private Bitmap foto;
 
-    public Usuario(String correo, String nombre, int cedula, Date fechaNacimiento, int telefono, String genero, Direccion direccion) {
+    public Usuario(String correo, String nombre, long cedula, Date fechaNacimiento, long telefono, String genero, Direccion direccion) {
         this.correo = correo;
         this.nombre = nombre;
         this.cedula = cedula;
@@ -53,11 +53,11 @@ public abstract class Usuario implements Serializable {
         this.nombre = nombre;
     }
 
-    public int getCedula() {
+    public long getCedula() {
         return cedula;
     }
 
-    public void setCedula(int cedula) {
+    public void setCedula(long cedula) {
         this.cedula = cedula;
     }
 
@@ -69,11 +69,11 @@ public abstract class Usuario implements Serializable {
         this.fechaNacimiento = fechaNacimiento;
     }
 
-    public int getTelefono() {
+    public long getTelefono() {
         return telefono;
     }
 
-    public void setTelefono(int telefono) {
+    public void setTelefono(long telefono) {
         this.telefono = telefono;
     }
